On branch master
Changes to be committed:
      modified:   README.md
      modified:   app.py

Changes not staged for commit:
      deleted:    dcmodels/datacs.py
      deleted:    xbi1_dcbot/__init__.py

There will be some radical changes in the basic functioning of the bot. It is being transformed into a personal service rather than a general public use. I think those who want to use this project will have to integrate the codes into their own bots after this commit.